+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 # Changelog - FeedHenry Javascript SDK
 
+## 2.1.4 - 2014-07-31 - IR233 - Niall Donnelly
+
+* 7656 - Refactored photo capture to return file URI instead of base64 string.
+
 ## 2.1.3 - 2014-07-17 - IR232 - Niall Donnelly
 
 7637 - Updated rules engine related to ticket 7637